Fix packet size calculation
Without this change, the size of the first packet was used as the size for all packets.
Since all iperf3 test packets are the same size, this does not change anything currently. However this may not have been the case in the future when packets may possibly have different sizes because of encapsulation.
Also delete some variables once they are no longer used, which would have made this error obvious.