Cut off the end of packet captures
The last part contains a connection teardown and pings, which are not supposed to be part of the measurement. This heuristic cuts off enough packets at the end, such that the part, which is kept is at most one second longer than the test duration. The last partial second is cut off by `analysis/statistics/statistics.py`.
Only cut off the packet capture from before the router, since the packets are in order there. The packets that are cut off before the router but are still visible after the router are ignored and don't influence statistics like e.g. packet loss.