Format LATEX debug output nicer
Tell `latexmk` to not wrap the lines of the log output.
Unfortunately it's not possible to never wrap any lines.
Increasing the value to 20000 results in the most unhelpful error message ever ("3 is a bad bad") if you can even find the error message among all the other debug output in the log.
See https://tex.stackexchange.com/questions/706447/why-am-i-getting-a-3-is-a-bad-bad-error.