Add possibility to rerun specific measurements
If there turns out to be a problem with some measurements but not all, this allows me to fix it without rerunning all tests.
Implementing it now instead of when I find a problem means that I don't have to later be careful to not change the test script for the tests that I don't want to rerun.